updated app.js to reflect current logged in user
